// src/features/hr/types/compensation.ts
export interface Compensation {
  id?: string;
  employee?: string;
  employee_number?: string;
  surname?: string;
  firstname?: string;
  position?: string;
  grade?: string;
  location?: string;
  project?: string;
  hire_date?: string;
  basic?: number;
  housing?: number;
  transport?: number;
  meal?: number;
  miscellaneous?: number;
  total_allowance?: number;
  thirteenth_month?: number;
  gross_total?: number;
}